requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.1
django-appconf==1.0.4
django-ckeditor==5.9.0
django-compressor==2.4
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-render-partial==0.4
django-taggit==1.3.0
mysqlclient==2.0.3
phonenumbers==8.12.8
pilkit==2.0
Pillow==8.1.0
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.3.1
